Added redux
