+++ package.json
@@ -10,6 +10,7 @@
     "react": "~17.0.2",
     "react-datepicker": "~4.1.1",
     "react-dom": "~17.0.2",
+    "react-redux": "^7.2.4",
     "react-scripts": "~4.0.3",
     "web-vitals": "~1.0.1"
   },
