+++ src/constants/actionTypes.js
@@ -1,0 +1,3 @@
+export const ADD_RECORD = "ADD_RECORD";
+export const DELETE_RECORD = "DELETE_RECORD";
+export const EDIT_RECORD = "EDIT_RECORD";
