+++ src/containers/Form/components/Table.js
@@ -1,54 +1,37 @@
-import React from 'react'
-import DataRow from './DataRow.js'
+import { connect } from 'react-redux';
 
-class Table extends React.Component {
-	constructor(props) {
-		super(props)
-		this.state = {
-			userData: [],
-			rows: []
-		}
-	}
+import { getRecordList } from '../../../selectors'
+import DataRow from './DataRow.js';
 
-	static getDerivedStateFromProps(nextProps, prevState) {
-		if (nextProps.userData !== prevState.userData) {
-			const dataRows = []
-			JSON.parse(nextProps.userData).forEach((data, i) => {
-				dataRows.push(
+function Table(props) {
+	return (
+		<table className="table">
+			<thead>
+				<tr>
+					<th>Name</th>
+					<th>Email</th>
+					<th>Date of Birth</th>
+					<th>Gender</th>
+					<th>Education</th>
+					<th>Picture</th>
+				</tr>
+			</thead>
+			<tbody>
+				{props.dataRows.map((data, index) => (
 					<DataRow
-						key={i}
-						name={data.input}
-						email={data.email}
-						dob={data.selectedDate}
-						gender={data.selectedRadioValue}
-						edu={data.selectedValue}
-						pic={data.selectedFile}
-						handleAction={(e) => nextProps.handleAction(e, i)}
+						key={data.id}
+						record={data}
+						handleEdit={props.handleEdit}
 					/>
-				)
-			});
-			return { rows: dataRows, userData: nextProps.userData }
-		}
-		else return null
-	}
-
-	render() {
-		return (
-			<table className="table">
-				<thead>
-					<tr>
-						<th>Name</th>
-						<th>Email</th>
-						<th>Date of Birth</th>
-						<th>Gender</th>
-						<th>Education</th>
-						<th>Picture</th>
-					</tr>
-				</thead>
-				<tbody>{this.state.rows}</tbody>
-			</table>
-		)
-	}
+				))}
+			</tbody>
+		</table>
+	)
 }
 
-export default Table
+const mapStateToProps = state => {
+	const dataRows = getRecordList(state)
+	return { dataRows }
+}
+
+export default connect(mapStateToProps, null)(Table)
