+++ src/reducers/index.js
@@ -1,0 +1,4 @@
+import { combineReducers } from "redux";
+import userData from "./userData.js";
+
+export default combineReducers({ userData });
